# Restaurant/Server/Restaurant/swagger/swagger.json
{"swagger":"2.0","info":{"title":"The RestaurantType API","description":"RestaurantType API with GOA","version":"1.0"},"host":"localhost:8080","basePath":"/V1","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/restauranttypes/":{"get":{"tags":["restauranttypes"],"summary":"list restauranttypes","operationId":"restauranttypes#list","produces":["application/vnd.restauranttype+json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestauranttypeCollection"}}},"schemes":["http"]},"put":{"tags":["restauranttypes"],"summary":"update restauranttypes","operationId":"restauranttypes#update","produces":["text/plain"],"parameters":[{"name":"payload","in":"body","description":"Restaurant Type","required":true,"schema":{"$ref":"#/definitions/UpdateRestauranttypesPayload"}}],"responses":{"200":{"description":"OK"},"404":{"description":"Not Found"}},"schemes":["http"]},"post":{"tags":["restauranttypes"],"summary":"create restauranttypes","operationId":"restauranttypes#create","parameters":[{"name":"payload","in":"body","description":"Restaurant Type","required":true,"schema":{"$ref":"#/definitions/CreateRestauranttypesPayload"}}],"responses":{"201":{"description":"Created"}},"schemes":["http"]}},"/restauranttypes/{id}":{"get":{"tags":["restauranttypes"],"summary":"get restauranttypes","operationId":"restauranttypes#get","produces":["application/vnd.restauranttype+json"],"parameters":[{"name":"id","in":"path","description":"Restaurant Type Index","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Restauranttype"}},"404":{"description":"Not Found"}},"schemes":["http"]},"delete":{"tags":["restauranttypes"],"summary":"delete restauranttypes","operationId":"restauranttypes#delete","produces":["text/plain"],"parameters":[{"name":"id","in":"path","description":"Restaurant Type Index","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"404":{"description":"Not Found"}},"schemes":["http"]}}},"definitions":{"CreateRestauranttypesPayload":{"title":"CreateRestauranttypesPayload","type":"object","properties":{"id":{"type":"string","description":"Restaurant Type Id","example":"Dolorem id corrupti atque."},"name":{"type":"string","description":"Restaurant Type Name","example":"Deleniti placeat voluptatem."}},"description":"Restaurant Type","example":{"id":"Dolorem id corrupti atque.","name":"Deleniti placeat voluptatem."},"required":["id","name"]},"Restauranttype":{"title":"Mediatype identifier: application/vnd.restauranttype+json; view=default","type":"object","properties":{"id":{"type":"string","description":"Restaurant Type Id","example":"Dolorem id corrupti atque."},"name":{"type":"string","description":"Restaurant Type Name","example":"Deleniti placeat voluptatem."}},"description":"Restaurant Type (default view)","example":{"id":"Dolorem id corrupti atque.","name":"Deleniti placeat voluptatem."},"required":["id","name"]},"RestauranttypeCollection":{"title":"Mediatype identifier: application/vnd.restauranttype+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Restauranttype"},"description":"RestauranttypeCollection is the media type for an array of Restauranttype (default view)","example":[{"id":"Dolorem id corrupti atque.","name":"Deleniti placeat voluptatem."}]},"UpdateRestauranttypesPayload":{"title":"UpdateRestauranttypesPayload","type":"object","properties":{"id":{"type":"string","description":"Restaurant Type Id","example":"Dolorem id corrupti atque."},"name":{"type":"string","description":"Restaurant Type Name","example":"Deleniti placeat voluptatem."}},"description":"Restaurant Type","example":{"id":"Dolorem id corrupti atque.","name":"Deleniti placeat voluptatem."},"required":["id","name"]}},"responses":{"Created":{"description":"Created"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK"}}}